% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospitalizationInSymptomaticPatientsValidation.R
\docType{package}
\name{HospitalizationInSymptomaticPatientsValidation}
\alias{HospitalizationInSymptomaticPatientsValidation}
\title{HospitalizationInSymptomaticPatientsValidation}
\description{
HospitalizationInSymptomaticPatientsValidation
}
